+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Unreleased
 - Match empty books for update using ISBN/ASIN even if OverDrive Link Integration is not enabled
 - Add configuration option to turn off updating of existing empty books
 - Improve identifying titles already in library by also matching ISBN/ASIN
+- Fix error if loan was sent to Kindle
 - Make UI more keyboard navigable
 
 Version 0.1.5 - 2023-07-15